LU-12208 tests: fix interop with old clients
We are unable to run the tests on the setup with mixed
clients versions installed because the clients older than
2.8.54 do not have set_hostid() added by LU-7134.

The hostid is to be set on all sever nodes including the
facets falover nodes. This is not done currently, and tests
fail on the setup like the following: ost1_HOST=host1
ost1failover_HOST=host2.

Patch fixes formatall() to call set_hostid() on server nodes
only, including the failover nodes.

Test-Parameters: trivial
Fixes: 7b654fc673 ("LU-7134 utils: Ensure hostid set for ZFS during mkfs")
Signed-off-by: Elena Gryaznova <c17455@cray.com>
Cray-bug-id: LUS-7242
Reviewed-by: Vladimir Saveliev <c17830@cray.com>
Reviewed-by: Alexander Boyko <c17825@cray.com>
Change-Id: I56f36c4b4e6cdadcb9940abfc9d2da8183b52509
Reviewed-on: https://review.whamcloud.com/34725
Tested-by: Jenkins
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Nathaniel Clark <nclark@whamcloud.com>
Reviewed-by: Alexandr Boyko <c17825@cray.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>