LU-12341 tests: Add kmemleak awareness to test-framework
If active kmemleak is detected, perform a clear operation to
ensure all non-Lustre related leaks are not getting in the way.

When it comes time to unload modules, first perform a scan
and then save the output if it's not empty, print to
syslog (for simplicity).
Also save /proc/modules content for the next step (we can save ogdb
from /tmp, but that seems to be getting stale and needs its own
fixing)

After modules unload perform another scan and if the result is non-empty,
output the saved /proc/modules output and the updated memleaks
into syslog as well

Test-Parameters: trivial
Change-Id: Ibba9047e4d8b98e7ab74aeb0906078549029ad43
Signed-off-by: Oleg Drokin <green@whamcloud.com>
Reviewed-on: https://review.whamcloud.com/34959
Tested-by: Jenkins
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Petros Koutoupis <pkoutoupis@cray.com>
Reviewed-by: James Simmons <uja.ornl@yahoo.com>