LU-4322 tests: re-enable 101a in dne config
We should re-enable 101a in the dne config, and also make
it more strict on discards.  This test should normally
result in 0 discards, because every page brought in by
readahead is used.

It is possible for randomness with the reads to lead to a
few discards, but no more than that.

Test-Parameters: trivial testlist=sanity,sanity,sanity,sanity
Signed-off-by: Patrick Farrell <pfarrell@whamcloud.com>
Change-Id: I53c6263c33c27d36b746c8fc56c8deebb4b713c7
Reviewed-on: https://review.whamcloud.com/35027
Reviewed-by: James Nunez <jnunez@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Oleg Drokin <green@whamcloud.com>