LU-11546 tests: enable large_dir support for tests
Enable the ldiskfs large_dir feature by default for all test
filesystems, so that we can verify it is not causing any issues
in regular testing.  If this is successful, it can be enabled
in mkfs.lustre permanently.

Existing conf-sanity.sh test_110() and test_111() exercise this
feature specifically, this patch is to ensure other tests have
no problems with this feature enabled.  There are other problems
with test_110() so it cannot be enabled at this time.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Ifff578551b57b05753fc10abb2d5294730254035
Reviewed-on: https://review.whamcloud.com/35358
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Artem Blagodarenko <artem.blagodarenko@gmail.com>
Reviewed-by: James Nunez <jnunez@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>